% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_ar1.R
\name{covar_ar1}
\alias{covar_ar1}
\title{Autoregression covariance matrix}
\usage{
covar_ar1(q, rho = 0.7)
}
\arguments{
\item{q}{dimension of covariance matrix (positive integer)}

\item{rho}{autoregression parameter (0 <= \code{rho} <= 1)}
}
\value{
Returns an AR(1) covariance matrix of parameter \eqn{\rho} and
dimension \eqn{q x q}.
}
\description{
Constructs a first order autoregression [AR(1)] covariance matrix.
}
\references{
\insertRef{MRCE}{tsmvrdata}
}
