% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_fgn.R
\name{covar_fgn}
\alias{covar_fgn}
\title{Fractional gaussian noise covariance matrix}
\usage{
covar_fgn(q, h = 0.9)
}
\arguments{
\item{q}{dimension of covariance matrix (positive integer)}

\item{h}{Hurst parameter (0 < \code{h} < 1)}
}
\value{
Returns an FGN covariance matrix with Hurst parameter
\code{h}.
}
\description{
Constructs a fractional gaussian noise (FGN) covariance matrix.
}
\references{
\insertRef{MRCE}{tsmvr}
}
