% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_curve.R
\name{error_curve}
\alias{error_curve}
\title{Calculate error curve}
\usage{
error_curve(Hat.list, Star = NULL)
}
\arguments{
\item{Hat.list}{a list of matrices (each dimension m-by-n)
representing the iterates of an optimization algorithm}

\item{Star}{matrix (m-by-n) representing the ground truth (optional)}
}
\value{
An error curve (numeric vector).
}
\description{
Calculates the error curve for a list of matrices
}
\details{
Given a list of matrices representing iterates of some optimization
algorithm and an optional "groud truth" matrix \code{Star}, this
function calculates the "error curve", which is the Frobenius norm
between each iterate and (1) the final iterate (if \code{Star} is
not specified) or (2) the ground truth (if \code{Star} is specified).
}
\note{
See also \code{\link[tsmvr]{tsmvr_solve}} and
\code{\link{plot_error_curve}}.
}
