% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate a sparse model}
\usage{
evaluate_model(Star, Hat, Sigma = NULL)
}
\arguments{
\item{Star}{matrix of true values (m-by-n)}

\item{Hat}{matrix of fitted values (m-by-n)}

\item{Sigma}{positive definite covariance matrix (m-by-m)}
}
\value{
A list of evaluation metrics, including: \cr \cr
\code{se}, squared error \cr
\code{me}, model error \cr
\code{tp}, number of true positives \cr
\code{tn}, number of true negatives \cr
\code{fp}, number of false positives \cr
\code{fn}, number of false negatives \cr
\code{tpr}, true positive rate \cr
\code{tnr}, true negative rate \cr
\code{acc}, accuracy \cr
\code{auc}, Area Under Curve \cr
\code{f1}, F1 score \cr
}
\description{
Calculates evaluation metrics between estimated and ground truth
sparse models (i.e., matrices)
}
\note{
Also see \code{\link{model_error}}.
}
