% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Make a synthetic dataset}
\usage{
make_data(n, p, q, b1 = sqrt(0.1), b2 = sqrt(0.1), sigma,
  rho_x = 0.6, type = "AR1", rho_err = 0.7, h = 0.9, power = 1,
  n_edge = 1, zero_appeal = 1, min_ev = 0.18, reps = 1,
  seed = NULL)
}
\arguments{
\item{n}{number of observations (positive integer)}

\item{p}{number of regressor features (positive integer)}

\item{q}{number of responses (positive integer)}

\item{b1}{Bernoulli parameter for controlling regressor matrix sparsity (positive integer)}

\item{b2}{another Bernoulli parameter for controlling regressor matrix sparsity (positive integer)}

\item{sigma}{scale of error term (positive numeric)}

\item{rho_x}{autoregression parameter for design matrix covariance matrix (0 < \code{rho} < 1)}

\item{type}{type of covariance matrix (string: 'AR1', 'FGN', or 'SFN')}

\item{rho_err}{autoregression parameter for AR(1) covariance matrix (0 < \code{rho} < 1)}

\item{h}{Hurst parameter for FGN covariance matrix (0 < \code{h} < 1)}

\item{power}{scaling power for SFN covariance matrix (positive numeric)}

\item{n_edge}{Barabasi algorithm number of edges per step for SFN covariance matrix (positive integer)}

\item{zero_appeal}{Barabasi algorithm baseline attractiveness for SFN covariance matrix (positive numeric)}

\item{min_ev}{minimum eigenvalue of SFN covariance matrix (\code{min_ev} > 0) (ensures matrix is PSD) )}

\item{reps}{number of randomly drawn datasets to return (positive integer)}

\item{seed}{seed for pseudorandom number generator}
}
\value{
Returns a list of length \code{reps}. Each entry is itself
a comprising a synthetic sparse multivariate dataset: a regressor
matrix \code{B}, a design matrix \code{X}, and a response matrix
\code{Y}. \code{B} has an expected sparsity of \code{b1} \code{x}
\code{b2}.

See also \code{\link{regressor_matrix}},
\code{\link{covariance_matrix}}, and
\code{\link[tsmvr]{tsmvr_solve}}.
}
\description{
Creates a random synthetic dataset for sparse multivariate
regression according to the model:
\deqn{Y = BX + E,}
where \eqn{X} is the design matrix, \eqn{B} is the regressor
matrix, \eqn{Y} is the response matrix, and \eqn{E} is the matrix
error term.
}
\references{
\insertRef{MRCE}{tsmvrextras}
}
