% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrnorm.R
\name{mvrnorm}
\alias{mvrnorm}
\title{Multivariate normal random vector generator}
\usage{
mvrnorm(n, mu = 0, Sigma, reps = 1)
}
\arguments{
\item{n}{number of realizations}

\item{mu}{mean (\eqn{\mu})}

\item{Sigma}{positive definite covariance matrix (\eqn{\Sigma})}
}
\value{
Returns a matrix of \code{n} stacked vectors, each an iid. realization of \eqn{N(\mu,\Sigma)}.
}
\description{
Generates vectors iid. from the multivariate normal distribution.
}
