% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_error_curve.R
\name{plot_error_curve}
\alias{plot_error_curve}
\title{Plot error curves}
\usage{
plot_error_curve(hat, star, up = 1, low = -5, left = 0,
  right = 100, auto = T)
}
\arguments{
\item{hat}{numeric vector representing the errors between iterates
and the final iterate}

\item{up}{upper limit of vertical axis}

\item{low}{lower limit of vertical axis}

\item{left}{left limit of horizontal axis}

\item{right}{right limit of horizontal axis}

\item{auto}{calculate limit automatically (default FALSE; if TRUE axes limits are ignored)}

\item{true}{numeric vector representing the errors between iterates
and the ground trush}
}
\value{
A ggplot object.
}
\description{
Plots the estimated and true error curves for iterates of an
optimization algorithm using ggplot2.
}
\note{
See also \code{\link[tsmvr]{tsmvr_solve}} and
\code{\link{error_curve}}.
}
