% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_hub.R
\name{precision_hub}
\alias{precision_hub}
\title{Precision matrix of scale free network}
\usage{
precision_hub(q, g, diag_val = 2, edge_val = 0.3)
}
\arguments{
\item{q}{square dimension of precision matrix (positive numeric integer)}

\item{g}{number of hubs, (numeric integer st. 1 <= g <= q)}

\item{diag_val}{values of diagonal entries HUB graph precision matrix (non-negative numeric)}

\item{edge_val}{values of HUB graph network edges}
}
\value{
Returns an Hub graph precision matrix
}
\description{
Constructs a Hub network precision matrix
}
\references{
\insertRef{chen2016high}{tsmvr}
}
