% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_sfn.R
\name{precision_sfn}
\alias{precision_sfn}
\title{Precision matrix of scale free network}
\usage{
precision_sfn(q, power = 1, n_edge = 1, zero_appeal = 1,
  min_ev = 0.18)
}
\arguments{
\item{q}{square dimension of covariance matrix (positive integer)}

\item{power}{power of the preferential attachment, (\code{a} > 0, default 1 or linear preferential attachment)}

\item{n_edge}{number of edges to add to the network per growth iteration (\code{n_edge} > 0, default 1)}

\item{zero_appeal}{the ‘attractiveness’ of the vertices with no adjacent edges (\code{b} >0 0, default 1)}

\item{min_ev}{minimum eigenvalue of resulting precision matrix (\code{min_ev} > 0, default 0.18)}
}
\value{
Returns an FGN covariance matrix.
}
\description{
Constructs a scale free network (SFN) precision matrix according
to the Barabasi algorithm.
}
\note{
See also \code{\link[igraph]{get.adjacency}},
\code{\link[igraph]{barabasi.game}}, and
\code{\link[igraph]{sample_pa}}.
}
\references{
\insertRef{chen2016high}{tsmvr}

\insertRef{barabasi}{tsmvrextras}
}
