% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbern.R
\name{rbern}
\alias{rbern}
\title{Bernoulli random number generator.}
\usage{
rbern(n, p)
}
\arguments{
\item{n}{number of draws}

\item{p}{Bernoulli parameter}
}
\value{
Returns a vector of length \code{n} with each entry drawn
iid. from Bern(\eqn{p}).
}
\description{
Random number generation from the Bernoulli distribution.
}
