% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared_error.R
\name{squared_error}
\alias{squared_error}
\title{Squared error}
\usage{
squared_error(U, V)
}
\arguments{
\item{U}{matrix (m-by-n)}

\item{V}{matrix (m-by-n)}
}
\description{
This function calculates the normalized squared error between two
matrices \eqn{U} and \eqn{V} according to the formula
\deqn{Trace [ (U-V)' (U-V) ]/(mn),}
where \eqn{m}, \eqn{n} are the dimensions of \eqn{U} (\eqn{V}).
}
\references{
\insertRef{MRCE}{tsmvr}
}
