% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_matrix.R
\name{covariance_matrix}
\alias{covariance_matrix}
\title{Construct covariance and precision matrices}
\usage{
covariance_matrix(q, type = "AR1", rho = 0.7, h = 0.9, power = 1,
  zero_appeal = 1, n_edge = 1, min_ev = 0.18)
}
\arguments{
\item{q}{dimension of covariance matrix (positive integer)}

\item{type}{type of covariance matrix (string: 'AR1', 'FGN', or 'SFN')}

\item{rho}{autoregression parameter for AR(1) covariance matrix (0 < \code{rho_err} < 1)}

\item{h}{Hurst parameter for FGN covariance matrix (0 < \code{H} < 1)}

\item{power}{scaling power for SFN covariance matrix (positive numeric)}

\item{zero_appeal}{Barabasi algorithm baseline attractiveness for SFN covariance matrix (positive numeric)}

\item{n_edge}{Barabasi algorithm number of edges per step for SFN covariance matrix (positive integer)}

\item{min_ev}{minimum eigenvalue of SFN covariance matrix (\code{min_ev} > 0) (ensures matrix is PSD) )}
}
\value{
A list of two matrices, the \code{covariance} matrix and the
\code{precision} matrix.
}
\description{
Constructs a covariance matrix and its associated precision matrix
of the following types: first-order autoregressive [AR(1)],
fractional gaussian noise [FGN], or scale free network [SFN].
}
\references{
\insertRef{MRCE}{tsmvrextras}

\insertRef{chen2016high}{tsmvr}
}
