% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrnorm.R
\name{mvrnorm}
\alias{mvrnorm}
\title{Multivariate normal random vector generator}
\usage{
mvrnorm(n, mu = 0, Sigma, reps = 1, seed = NULL)
}
\arguments{
\item{n}{number of realizations}

\item{mu}{mean (\eqn{\mu})}

\item{Sigma}{positive definite covariance matrix (\eqn{\Sigma})}

\item{reps}{number of random matrix realizations to return}
}
\value{
Returns a \code{rep} length list of matrices of \code{n}
stacked vectors, each an iid. realization of \eqn{N(\mu,\Sigma)}.
If \code{reps}=1, then a single matrix returned not in a list. Otherwise,
\code{reps} matrices are returned in a list.
}
\description{
Generates vectors iid. from the multivariate normal distribution.
}
