% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor_matrix.R
\name{regressor_matrix}
\alias{regressor_matrix}
\title{Sparse regressor matrix}
\usage{
regressor_matrix(p, q, b1 = sqrt(0.1), b2 = sqrt(0.1), seed = NULL)
}
\arguments{
\item{p}{number of regression features (positive integer)}

\item{q}{number of regression responses (positive integer)}

\item{b1}{a Bernoulli parameter controlling the sparsity of the regressor matrix (0 <= \code{s1} <= 1)}

\item{b2}{a second Bernoulli parameter controlling the sparsity of the regressor matrix (0 <= \code{s2} <= 1)}

\item{seed}{seed for pseudo-random number generator (numeric)}
}
\value{
Returns a random sparse regressor matrix of dimension
\code{p} x \code{q} with expected sparsity \code{b1} \eqn{x} \code{b2}.
}
\description{
Creates a random realization of a sparse multivariate regressor
matrix.
}
\references{
\insertRef{MRCE}{tsmvrextras}
}
